+++ github-runner-manager/src/github_runner_manager/reactive/consumer.py
@@ -27,6 +27,9 @@ logger = logging.getLogger(__name__)
 
 Labels = set[str]
 
+PROCESS_COUNT_HEADER_NAME = "X-Process-Count"
+WAIT_TIME_IN_SEC = 60
+RETRY_LIMIT = 5
 # This control message is for testing. The reactive process will stop consuming messages
 # when the message is sent. This message does not come from the router.
 END_PROCESSING_PAYLOAD = "__END__"
@@ -125,8 +128,31 @@ def consume(  # noqa: C901
                 if msg.payload == END_PROCESSING_PAYLOAD:
                     msg.ack()
                     break
+
+                msg.headers[PROCESS_COUNT_HEADER_NAME] = (
+                    msg.headers.get(PROCESS_COUNT_HEADER_NAME, 0) + 1
+                )
+                msg_process_count = msg.headers[PROCESS_COUNT_HEADER_NAME]
+
                 job_details = _parse_job_details(msg)
                 logger.info("Received reactive job: %s", job_details)
+
+                if msg_process_count > RETRY_LIMIT:
+                    logger.warning(
+                        "Retry limit reach for job %s with labels: %s",
+                        job_details.url,
+                        job_details.labels,
+                    )
+                    msg.reject(requeue=False)
+                    continue
+
+                if msg_process_count > 1:
+                    logger.info(
+                        "Pause job %s with retry count %s", job_details.url, msg_process_count
+                    )
+                    # Avoid rapid retrying to prevent overloading services, e.g., OpenStack API.
+                    sleep(WAIT_TIME_IN_SEC)
+
                 if not _validate_labels(
                     labels=job_details.labels, supported_labels=supported_labels
                 ):
@@ -248,7 +274,7 @@ def _spawn_runner(
     logger.info("Reactive runner spawned %s", instance_ids)
 
     for _ in range(5):
-        sleep(60)
+        sleep(WAIT_TIME_IN_SEC)
         logger.info("Checking if job picked up for reactive runner %s", instance_ids)
         if platform_provider.check_job_been_picked_up(metadata=metadata, job_url=job_url):
             logger.info("Job picked %s. reactive runner ok %s", job_url, instance_ids)