+++ docs/changelog.md
@@ -2,7 +2,20 @@
 
 This changelog documents user-relevant changes to the GitHub runner charm.
 
-## 2025-07-15
+## 2025-07-24
+
+- Fix an issue with infinite retry of a reactive job message.
+
+## 2025-07-22
+
+- Removed support for using both jobmaanger and GitHub at the same time.
+
+
+## 2025-07-18
+
+- Fix an issue where flushing runners does not include reactive processes. This cause some reactive runner to spawn with old code after upgrade.
+
+## 2025-07-16
 
 - Split the reconcile_duration buckets for Prometheus metrics into a larger bucket set.
 - Fix the incorrect default value of the aproxy-exclude-addresses configuration.
@@ -13,6 +26,7 @@ This changelog documents user-relevant changes to the GitHub runner charm.
  due to a bug on the OpenStack side: https://bugs.launchpad.net/nova/+bug/2095364
 
 ### 2025-06-30
+
 - New configuration options aproxy-exclude-addresses and aproxy-redirect-ports for allowing aproxy to redirect arbitrary TCP traffic
 - Added prometheus metrics to the GitHub runner manager application.